refactor(vault): share splitLines / joinLinesLf across the document readers + serializers
The maps.txt and map.msg round-trip readers had an identical line-splitter, and both serializers an
identical LF-join — enough duplicated lines to trip SonarCloud's new-code duplication gate (4% > 3%).
Hoist both into geck::text (reader/TextParsing.h): splitLines(content, finalNewline) strips a trailing
'\r' and reports the trailing-newline state; joinLinesLf(lines, finalNewline) is its inverse. Behaviour
is unchanged; 47 round-trip assertions still pass.
